@@ -24,6 +24,8 @@
   "prettier": "@mizdra/prettier-config-mizdra",
   "renovate": "github>mizdra/renovate-config-mizdra",
   "devDependencies": {
+    "@babel/core": "~7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
     "@jest/types": "^27.5.1",
     "@mizdra/eslint-config-mizdra": "^1.1.0",
     "@mizdra/prettier-config-mizdra": "^0.4.0",
@@ -36,6 +38,7 @@
     "@types/yargs": "^16.0.4",
     "@typescript-eslint/eslint-plugin": "^5.16.0",
     "@typescript-eslint/parser": "^5.16.0",
+    "babel-jest": "^28.1.0",
     "enhanced-resolve": "^5.8.3",
     "eslint": "^8.11.0",
     "eslint-config-prettier": "^8.5.0",
@@ -45,9 +48,8 @@
     "npm-run-all": "^4.1.5",
     "prettier": "2.6.0",
     "strip-ansi": "^7.0.1",
-    "ts-jest": "^27.1.4",
     "tsc-watch": "^4.2.9",
-    "typescript": "^4.6.0-dev.20220211"
+    "typescript": "^4.7.2"
   },
   "dependencies": {
     "boxen": "^6.2.1",
@@ -66,6 +68,9 @@
   },
   "peerDependencies": {
     "eslint": "^7.0.0 || ^8.0.0"
+  },
+  "resolutions": {
+    "@babel/core": "~7.17.12"
   },
   "engines": {
     "node": "^14.13.1 || >=16.0.0"
