@@ -1,6 +1,5 @@
 import { join } from 'path';
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-import cachedir = require('cachedir');
+import cachedir from 'cachedir';
 import { VERSION } from '../cli/package.js';
 
 /**
