@@ -1,5 +1,6 @@
 declare module 'cachedir' {
   // eslint-disable-next-line import/no-default-export
   function cachedir(name: string): string;
-  export = cachedir;
+  // eslint-disable-next-line import/no-default-export
+  export default cachedir;
 }
